[Refactoring] 영구저장소에 저장하기 위해 SpeedInfo 모델의 Any타입이었던 value 프로퍼티를 Double로 수정 후 관련 코드들 정리 및 단순화

+++ SimpleLocationTracking/Extension/Double + Time Notation.swift
@@ -1,5 +1,5 @@
 //
-//  Int + Time Notation.swift
+//  Double + Time Notation.swift
 //  SimpleLocationTracking
 //
 //  Created by TAEHYOUNG KIM on 11/30/23.
@@ -7,12 +7,12 @@
 
 import Foundation
 
-extension Int {
+extension Double {
     /// hh:mm:ss
     var hhmmss: String {
-        let hours = self / 3600
-        let minutes = (self % 3600) / 60
-        let seconds = (self % 3600) % 60
+        let hours = Int(self) / 3600
+        let minutes = (Int(self) % 3600) / 60
+        let seconds = (Int(self) % 3600) % 60
 
         return String(format: "%02d:%02d:%02d", hours, minutes, seconds)
     }
@@ -23,9 +23,9 @@ extension Int {
     /// 1분
     /// 48초
     var resultTime: String {
-        let hours = self / 3600
-        let minutes = (self % 3600) / 60
-        let seconds = (self % 3600) % 60
+        let hours = Int(self) / 3600
+        let minutes = (Int(self) % 3600) / 60
+        let seconds = (Int(self) % 3600) % 60
 
         if hours > 0 && minutes > 0 {
             return "\(hours)시간 \(minutes)분"
